+++ ZenDevToolkit/URLEncoderView.swift
@@ -200,7 +200,6 @@ struct URLEncoderView: View {
                         .font(.system(size: 12, design: .monospaced))
                         .scrollContentBackground(.hidden)
                         .padding(12)
-                        .focused($isInputFocused)
                         .onChange(of: inputText) {
                             updateCharacterCount()
                             if !inputText.isEmpty {